Add `dummy.cxx` file to automake rules to force C++ linking
Automake will automatically use the C frontend to link together object
files unless there are obviously C++-style source files listed as
sources for the link step.  This can cause issues if the C frontend is
not smart enough to e.g. auto-link in the C++ standard library.  This
can be worked around by manually sneaking in a `dummy.cxx` file as a
fake dependency to the automake rules.  Note that no `dummy.cxx` file
actually exists, we just pretend that it is part of the list within
Automake.  See this page for more details:
https://www.gnu.org/software/automake/manual/html_node/Libtool-Convenience-Libraries.html