Remove unconditional `libstdc++` linkage
Compilers such as `clang` prefer to use `libc++` as their C++ standard
library implementation, as opposed to `libstdc++`.  By removing this
explicit linkage, we are able to let the compiler automatically include
whichever library it uses by default.  This is necessary on platforms
such as macOS and FreeBSD, where `libc++` via `clang` is the default.